# More than triple linter speed