--- a/lunatrace/bsl/common/package.json
+++ b/lunatrace/bsl/common/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lunatrace/lunatrace-common",
-  "main": "build/main/index.js",
+  "main": "src/index.ts",
   "typings": "build/main/index.d.ts",
   "license": "BSL-LunaTrace",
   "type": "commonjs",