--- a/lunatrace/bsl/frontend/tsconfig.json
+++ b/lunatrace/bsl/frontend/tsconfig.json
@@ -23,6 +23,10 @@
   "include": [
     "src"
   ],
+  "exclude": [
+    "node_modules",
+    "src/api/generated.ts"
+  ],
   "references": [
     {"path": "../common"}
   ]