--- a/lunatrace/bsl/logger/package.json
+++ b/lunatrace/bsl/logger/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lunatrace/logger",
-  "main": "build/main/index.js",
+  "main": "src/index.ts",
   "typings": "build/main/index.d.ts",
   "license": "BSL-LunaTrace",
   "type": "commonjs",