--- a/lunatrace/npm-package-cli/package.json
+++ b/lunatrace/npm-package-cli/package.json
@@ -8,7 +8,7 @@
   },
   "homepage": "https://github.com/lunasec-io/lunasec",
   "license": "Apache-2.0",
-  "main": "dist/index.js",
+  "main": "src/index.ts",
   "typings": "build/main/index.d.ts",
   "repository": {
     "type": "git",